Combined WebApiClient into RemoteTranslationEngine
